Starts porting CSP DR1 to OOD
