+++ sndata/__init__.py
@@ -5,9 +5,9 @@
 surveys.
 """
 
-from . import csp, des, essence, jla, sdss
-from ._combine_data import CombinedDataset
-from .exceptions import ObservedDataTypeError as _ObservedDataTypeError
+#from . import csp, des, essence, jla, sdss
+#from ._combine_data import CombinedDataset
+#from .exceptions import ObservedDataTypeError as _ObservedDataTypeError
 
 __version__ = '0.9.5'
 __author__ = 'Daniel Perrefort'
