+++ sndata/csp/data/tables/table3.dat
@@ -1,0 +1,78 @@
+CN 2004ey 16203 1307  9640  381  9327  279  9238  191  9942  200 11204 333
+CN 2005bo 11928  788  9527  153  8291  336  8527  118 10065  202 10857 291
+CN 2005el 12407   41  9668   46  8703   43  8966   29 10329  182 11003  51            1328  294
+CN 2005hc            10292  336  9071  587  9835  259 11038  489 11163 219
+CN 2005kc            10089  352  8960  429  8916   93  9962  217 10554 239             707  270
+CN 2005ki            10522  738                       10268    0  9811  92
+CN 2005ku            11334  313  9793  842  9972  297 11783 1592 12215 697
+CN 2005na             9852  114                       10903  525 10763  62             608   66
+CN 2006D              9684   42  9026  948  9216  196 10389   38 10867  27 12467   14  911  154
+CN 2006ax             8726  100  8360  272  8375   72  8897  207 10170  78            1190  250
+CN 2006et             9101  201                        9858  108 10123 223
+CN 2006fw 16447  839 11550  204 10360  419  9597  158            10819 358             811  628
+CN 2006is                       13046  298 15502 2581 13903  721 14122 218 14125  815 1064  230
+CN 2007A              9641   69  8956 1491  9037  322  9840  114 10990  41
+CN 2007bm 12442 1387  9409  154  8566  596  8733  124 10563  267 10817 440 11575  667
+CN 2007ca 15238 1190 10206  316  9802  216  9438  155 10174  129 11085 181
+CN 2008ar 17691  768  9586  101  8644  164  8948   20 10640   97 10471  54 10579 2602 1007  129
+CN 2008bc 17277  340 10503   66  9929   78 10139   65 12002  176 11547  79 12781  284  953  256
+CN 2008bf 12838  153 10565   32  9995   65  9981   35 10762   67 11491 103  9866  545  905  130
+CN 2008bq 17890  931  9351  137  8711  366  9101  252 10021  305 10786 219 20225  726 1232  337
+CN 2008fp            11075   23  9407    7  9785   53 10944  115 10986  37 12039   51 1007  367
+CN 2008hv 18288   38  9620   91  8457   48  9137   58 10543   53 10926  56 11155  323  813   67
+CN 2009D  18323   14  8824   59  8203   47  8532  110 10115   75 10243  57             682  479
+CN 2009aa             8764  216  8494   83  8340   80  9768   40 10546  28            1267  269
+CN 2009ab 11347  701  8783   12  8136   25  8756   31 10424  194 10563  22
+CL 2004eo 14068  186  8682   81  7724   56  8399   86  9854   32 10352 119 11751  128 1370  122
+CL 2004gs 12601  842  9175  254  7520  668  7949  338 10690   46 11483  90 13600 1009
+CL 2005bl 12787   37                                   9485  208  9930  86 11796   70 2266  177
+CL 2005ke 11472  724 10194 1525  7283   94  6575   55  9514   33  9811  92            1885  106
+CL 2006bd                                             10017  436 11207 368 13428  997
+CL 2006eq             8866 1632                       10927  414 11742 586
+CL 2006gj 11592  983  9645   98  7133  793  7913  328 10438   52 11256 107 11426  518 1839  350
+CL 2006gt 12444  108  9046   51  6335   76  7140  118  9490   72 10394 114 11660   91 1864  201
+CL 2006kf 15025 1382  9548   48  8895   43  9086   42 10634   26 11432  72            1721  203
+CL 2006mr 11950  916             8021  316  6446  340  9661   54 10213  82 13134  193 2604  240
+CL 2006os 16898  510 10755  292  8894  294  9326  375 10952  410 11911 620 13042  406 1968 1183
+CL 2007N                                              10301  229 11213 347
+CL 2007al 12038  494                                   9418  196  9850 144 12101  364 2050  310
+CL 2007ax 15905 1108             6847  344  6245  261  9563  131  9773 206 14357 1409
+CL 2007ba 12802 1496  8826  513  7870  948  8270  198  9996  185 10612 212 12099  459
+CL 2007bc 11877   95  8965   55  8075  286  8446   28  9986   35 10382  33 11211  591
+CL 2007hj            10292   98                       11167   56 12297  43
+CL 2007on 14675  938  9988  115  7819   34  8186   53 10482   49 11242  28 12369  123 1988  109
+CL 2007ux 12230  632  8933  263  6597  404  7306  184 10105  148 10743 318 11911  355
+CL 2008R  13382  581  9054  170  7281  310  7618  113 10204  173 10922 245            1635  252
+CL 2009F  13521  915  9059   60  8863  336  8247  221 10738  149 11004  53 11913  818
+BL 2004dt 10691  111 10381  208  9848   79  9076   30 12025   88 14799  96 13066  414 5047  103
+BL 2004ef            10286   85  8689  108  9802   84 11750   99 12039 133 14668 1453
+BL 2005A             10176 1014 12147  386 11577  210 11095  445 14976 279
+BL 2005W                                              10549  291 10867 468
+BL 2005ag 16357 1966 10691  288  8528  553 10385  438 10151  865 11558 370
+BL 2005am            10679   75                       11713   94 12157  70            1692   77
+BL 2006X             12750  127  9878  268 11515  192 12590   53 16232  42            3661  107
+BL 2006br            13236  197 11059  266 10473  435 12218 1810 14174 442 14343  900
+BL 2006ef             9784  136                       11396  723 12375 611
+BL 2006ej            11016  286                       12135  511 12249 373
+BL 2006ot            12876  225                                  13898 881
+BL 2007af 13842 1178  9301   30  8637   73  8444   33 10246   63 10918  24 11734   72 1414  552
+BL 2007as 20965 1095 11749  301 10288  546 10665  228 12460  616 12962 621 16392 1963
+BL 2007bd 16815 1179 11047  138  9963  296  9714    4 11591   96 12547  95 12789 1181 1496  873
+BL 2007jg 24293 1993  9796  211  9093  295  9604  350 10653  451 12615 394 21810 1001 2723  475
+BL 2007le 19850  210 11752   45 10997   66 10540   36 12135   63 12539  27 13381  225 1172  154
+BL 2007nq 15021  872 10404  199  8659  360  9696  120 11308  160 12111 289 11930  322 1618  351
+BL 2008gl            10262  328  8706  422  9577  226 10976  927 11989 525 12217  685
+BL 2008hu 14566  874 11023  277  9485  859 10026  166 12160  366 12343 671 12414  683
+BL 2008ia 10801  116  9923  126  9264   95  9154   16 10573   40 11182  45 12619  139  843   49
+BL 2009Y  21193 1087            12024  309 12293  106 12829  216 14876  48 13949   72 1713  262
+BL 2009ag             8138  145  7418  405  7922   76  9301  242 10261 183            1766  292
+SS 2004gu 19104  671             8597  822  9195  134            10631 302
+SS 2005M  14708  700  7886  201  6793   64  6952  122  8199  216  8761 181            1884  419
+SS 2005bg 12213  750  9984  130  9972  284  9017   99 10517  137 10793 283 12419  654 1135  537
+SS 2005eq             9685   51                       11081  212 10097 110             944  225
+SS 2005hj            11096  240  9175   62  9292  103 10377  206 10538  52 10678  451  167  531
+SS 2006hx 15838 2030  9304 1209  9650  269  9637  339  9653  883 10593 393
+SS 2007S              9253  105  8515  176  8776   50  8602   37 10031  56             824  391
+SS 2007ai 15562  771  8726  401  8092   28  7818  144  9053   43  9980 109
+SS 2008C  14187 1000             8036  656  7867  133  9628  405 10526 597 11361  261 1231  600
+SS 2009ad 11342 1047  9815   60  8999   56  9247  121  8235  192 10394  22
