+++ sndata/csp/data/tables/table6.dat
@@ -1,0 +1,78 @@
+CN 2004ey 107.2  5.3 13.6 1.5  96.8  1.5 129.9  3.2  91.1  3.3 14.2 1.1  99.7  3.5  65.0 16.5
+CN 2005bo            16.8 1.6  85.2  3.9 134.1  5.3  80.5  2.8 23.2 1.9  99.2  1.8 158.8  5.9
+CN 2005el 110.7  1.8 20.6 0.5  82.4  5.8 112.9  3.9  82.1  1.2 20.5 1.5  93.8  1.7 130.1  5.2
+CN 2005hc            10.3 5.2 102.0  5.2 125.1  1.1  89.3  3.8 18.3 1.5  84.3  1.6
+CN 2005kc                      75.7  2.4 130.7  5.2  75.4  1.3 25.8 1.0  91.2  1.1
+CN 2005ki 116.1  1.6 20.9 1.4  86.7  1.4 110.8  1.1  81.4  0.6 22.3 2.0 104.3  0.5  86.0  1.1
+CN 2005ku            11.5 1.7  79.7  8.1  91.0  3.7  54.6  1.4 12.1 1.3  83.0  3.0
+CN 2005na  87.7  6.2 11.0 0.5 116.7  4.8 117.1 15.4  89.6  2.4 16.3 2.0  78.7  3.8  91.4 22.5
+CN 2006D   96.8  1.2 20.9 0.2 182.9  0.5 136.1  5.3  87.3  3.0 27.2 2.3  98.5  0.7 188.0  9.7
+CN 2006ax 117.0  1.0 13.4 0.6  98.3  1.8 118.9  2.8  84.8  2.5 12.1 0.7  93.5  2.0 131.6 12.8
+CN 2006et 116.7  2.7 16.6 2.5 103.6  3.3 159.2  6.9  76.9  2.1 23.5 3.3  73.2  6.0 127.3  3.8
+CN 2006fw 108.9  5.4 19.4 2.5  97.9  4.8 106.2  2.4  49.6  1.5 13.2 1.6  76.2  3.8
+CN 2006is                                119.8  5.9  79.8  2.3  9.6 1.2  80.3  0.9 137.3  3.7
+CN 2007A   99.8  1.2 11.8 1.2  89.9  2.4 119.2  3.3  79.8  2.7 15.9 2.1  87.7  1.6 106.0  5.7
+CN 2007bm 111.1  5.8          129.7  5.4 149.3  2.9  70.9  1.4 27.3 0.7  95.0  1.4 160.1  3.6
+CN 2007ca 123.7  4.4 12.3 0.4  86.0  6.2 111.2  1.9  77.6  3.8 11.4 2.5  86.1  0.8 114.9 26.1
+CN 2008ar 166.1  1.5 17.4 0.6 110.1  4.2 149.5  2.0  96.6  2.5 22.2 1.0  99.3  1.9 179.9  5.0
+CN 2008bc 137.4  0.9 14.7 0.4 104.3  5.0 134.2  0.8  73.7  2.6 14.6 0.4  96.7  1.1 152.4  6.2
+CN 2008bf 129.5  1.3 11.3 0.3  97.9  5.3 113.7  2.1  84.7  1.8 13.4 0.7  78.8  2.1 106.7  7.3
+CN 2008bq            13.0 0.6  94.0  0.8 148.4  1.2  80.8  0.9 15.7 0.9  94.2  0.7 171.1  8.7
+CN 2008fp             5.2 1.2  94.9  0.7 110.4  0.5  71.9  1.1 11.6 0.8  74.1  1.8  85.1  5.8
+CN 2008hv            20.5 2.4  88.2  4.5 115.0  2.3  80.5  1.5 20.9 0.5  99.2  1.9 121.0  3.4
+CN 2009D  122.7  0.9 11.6 1.2  96.9  2.4 135.8  1.0  72.1  4.8 10.7 1.3  81.9  0.9 161.0  0.6
+CN 2009aa            20.8 3.7  88.3  1.1 137.8  0.9  88.0  1.4 24.5 1.1  95.6  1.2 131.2  7.7
+CN 2009ab            19.4 1.0 114.4  4.4 146.0  1.0  80.6  2.2 28.7 0.8  99.1  0.7 167.8  2.3
+CL 2004eo 127.7 16.3 32.2 2.6 142.7  1.6 165.8  6.9  82.1  0.8 37.1 2.5 117.4  5.7 192.7 15.4
+CL 2004gs 105.7  7.9 35.3 1.9 104.5  9.1 197.9  3.5  87.4  2.3 47.8 3.1 134.1  1.9 243.1 18.9
+CL 2005bl 119.0  3.9          277.2  2.3 183.3 14.2  29.3  3.4 47.9 1.8 106.7  1.8 296.3  3.5
+CL 2005ke 111.6  9.9          269.5  3.1 148.7  4.5  33.3  1.8 57.4 1.4 103.5  1.0 290.4 11.6
+CL 2006bd                     281.6 10.0 187.1  2.8  41.5  5.2 51.8 6.8 128.7  3.0 378.8 20.5
+CL 2006eq            48.6 3.0 201.7  6.4 195.1  4.9  76.6  5.7 56.7 4.5 145.3  5.1
+CL 2006gj 104.5  2.1 26.9 1.8 100.5  3.2 137.3  5.0  81.6  5.8 46.3 5.4 112.5  4.2 242.2 40.3
+CL 2006gt 153.1  5.5 20.9 0.3 179.0  0.2 139.0  6.3  72.7  2.9 56.8 3.1 121.2  2.1 267.7 14.3
+CL 2006kf 102.9  4.4 23.1 2.9  86.8  5.6 140.8  0.7  83.3  3.6 33.5 0.4 115.9  5.0 153.9  5.4
+CL 2006mr                     247.2  5.2 133.6 26.5  19.9  4.3 42.2 5.5 105.6  5.1 317.3 14.7
+CL 2006os            31.5 1.7 112.0  4.8 187.0  3.6  98.5  3.0 42.0 2.6 145.6  2.4 291.0  9.8
+CL 2007N                                 166.6  1.1  40.8  0.6 44.5 0.7  96.4  1.2 413.3  1.0
+CL 2007al 110.0 54.5          202.7 34.1 140.2 13.4  14.5  2.0 40.7 7.3  74.4 10.4 247.2 11.1
+CL 2007ax 123.4  3.4          228.8  4.6 122.0 12.4  26.1  3.0 43.6 1.9  85.9  7.7 444.4  6.5
+CL 2007ba 111.7  9.1 20.0 0.8 211.4  7.1 130.2  5.7  65.5  0.9 48.4 9.5 111.4  1.2 292.1  5.5
+CL 2007bc 120.1  3.6 24.2 0.3  96.5  5.0 154.7  2.0  89.5  2.2 33.5 1.3 101.3  0.5 161.9  7.8
+CL 2007hj 130.5  1.7 33.9 0.5 164.8  1.2 171.2  1.4  87.5  1.7 57.6 0.6 140.3  1.9 281.2  2.0
+CL 2007on 105.1  3.1 24.7 0.1 140.2  1.5 119.6  2.0  64.3  1.5 44.0 0.7 117.2  0.2 237.4 17.9
+CL 2007ux  91.2 25.3 28.5 5.9 142.4  9.6 149.5 12.2  68.5  3.6 52.8 6.0 120.2 10.6 248.9 26.2
+CL 2008R   95.7  1.6 21.5 1.2 143.2  6.7 147.9  9.9  70.1  1.1 46.3 0.9 120.4  2.0 225.5 16.0
+CL 2009F                      242.6  9.8 194.8  6.9  30.7  2.5 48.9 1.6 115.7  3.7 279.4  5.5
+BL 2004dt 140.4  8.0 24.3 0.5 111.6  2.8 118.8  2.7  72.6  0.9 20.2 0.9 172.7  1.9 134.9  5.8
+BL 2004ef 155.0  1.2 27.1 0.9 108.3  2.5 186.8  2.5  86.5  2.5 26.2 1.2 131.3  2.1 158.1 33.4
+BL 2005A  210.0 21.8 21.8 1.9 128.2  2.5 306.2 10.9  58.6 12.6 14.0 1.8 142.6  3.1 400.9 18.4
+BL 2005W                                             83.9  0.4 24.3 1.0 112.8  1.3 220.0  2.6
+BL 2005ag 119.9 17.5 11.1 1.1 103.5 11.3 133.1  6.0  92.0 11.8 14.2 2.9 106.1 17.9
+BL 2005am 117.5  2.6 24.4 0.8  87.3  7.9 169.8  2.4 104.8  3.1 25.6 1.7 121.5  0.6 199.0  3.0
+BL 2006X  173.3  4.9 25.8 2.0  93.7 12.9 241.5  1.2  89.2  1.6  7.8 1.1 183.2  0.6 283.6  2.4
+BL 2006br            22.7 2.7  77.0  2.7 118.8 26.2  75.4  6.0 14.2 4.1 121.4 12.3 133.3  6.5
+BL 2006ef 126.2  1.9 27.3 0.7 119.8  3.2 154.8  2.5 105.0  0.8 27.8 0.4 132.2  0.7 177.5  1.9
+BL 2006ej  94.2  2.5 30.8 1.1 123.6  3.4 165.1  1.9  92.4  1.1 30.1 0.3 118.2  0.4 194.6  7.7
+BL 2006ot 148.5 13.6 36.7 5.0 114.3 23.5 155.4 42.9  54.4 19.8 20.9 6.0 127.7 36.6
+BL 2007af 117.0  2.6 21.6 0.3 112.2  2.4 161.8  3.7  93.4  0.8 22.4 0.3 105.2  0.9 191.8  6.1
+BL 2007as 177.1  1.2 21.9 0.8 107.6  1.0 158.3  2.1  92.0  1.0 20.8 0.8 136.8  0.9 234.8  7.4
+BL 2007bd 119.3  1.6 20.4 0.5  91.8  2.5 125.4  2.2  90.5  1.5 16.5 1.1 113.5  1.2 130.9 14.9
+BL 2007jg 189.7  2.8 12.4 0.7  89.0 11.8 124.3 13.9  82.3  4.7 17.9 0.5 114.3 14.5 304.3  2.1
+BL 2007le 160.7  1.5 14.7 0.6  94.6  4.6 160.3  1.4  79.4  3.7 14.2 0.9 108.2  2.6 188.6  6.1
+BL 2007nq            24.2 1.1            129.2  3.2  82.2  1.8 29.8 1.1 114.6  1.6 170.0 11.0
+BL 2008gl            21.6 0.5  97.2  3.9 150.7  8.2  87.5  5.2 23.4 2.1 116.8  6.5 139.6  3.2
+BL 2008hu 107.1  3.5          133.7  4.8 158.0  3.6  95.8  1.8 25.9 3.6 128.1  1.6 221.1  5.1
+BL 2008ia 140.6  1.6 23.9 0.6  97.1 16.6 143.4  7.7  89.9  4.0 18.0 0.5 110.5  5.3 144.3  7.6
+BL 2009Y  152.9  3.2  7.6 0.6 111.3  2.9 213.0  4.0  77.5  2.9  8.8 1.2 141.1  0.8 165.2  4.2
+BL 2009ag            18.2 1.0 108.9  2.1 166.0  2.1  86.8  0.2 21.4 0.3 110.2  3.9 260.4  1.4
+SS 2004gu  69.3  4.2  3.0 1.0 103.7  4.9 128.6  3.0  72.8  1.9  4.5 1.2  67.2  1.8 179.4 27.6
+SS 2005M  100.4 12.5  9.6 1.4  94.9  6.2 141.7  6.0  71.5  3.0 12.0 3.2  54.7  4.0  59.3 23.8
+SS 2005bg             7.2 0.6  75.5  2.4  80.1  2.5  60.8  1.7 12.2 1.2  53.9  0.9  50.3  8.2
+SS 2005eq 104.2  3.4  6.9 0.4  97.2  1.6 116.6  5.0  62.3  3.1 12.3 1.3  56.6  1.2  79.3  4.4
+SS 2005hj                     111.0  3.0 103.3  3.7  70.2  4.0 13.5 1.2  58.5  3.3 132.3 12.7
+SS 2006hx                      81.3  4.0  89.9  3.6  55.4  2.6 20.6 1.6  54.3  2.2
+SS 2007S   91.6  2.2  7.0 0.9  93.1  1.8 131.8  3.1  68.7  1.5 11.4 0.5  59.4  0.2 105.3 12.3
+SS 2007ai  99.0  2.7  4.2 1.0 104.8  1.7 150.9  5.0  59.7  1.1 14.2 1.4  62.9  4.7  86.0  3.7
+SS 2008C                                 131.7  1.7  48.4  1.2 15.4 0.9  66.6  0.9
+SS 2009ad             9.8 0.4  94.2  3.7 106.3  3.2  68.1  1.0 13.8 1.6  68.8  0.9  57.0  8.5
